+++ llarm-asm/src/disassemble/generators/thumb/misc.cpp
@@ -18,7 +18,7 @@ using namespace internal;
  * 
  * reference: A7-24
  */
-std::string generators::thumb::misc::BKPT(const u16 code, const settings settings) {
+std::string generators::thumb::misc::BKPT(const u32 code, const settings settings) {
     const u8 immed_8 = llarm::util::bit_range<u8>(code, 0, 7);
 
     return util::make_string(
@@ -36,7 +36,7 @@ std::string generators::thumb::misc::BKPT(const u16 code, const settings setting
  * 
  * reference: A7-102
  */
-std::string generators::thumb::misc::SWI(const u16 code, const settings settings) {
+std::string generators::thumb::misc::SWI(const u32 code, const settings settings) {
     const u8 immed_8 = llarm::util::bit_range<u8>(code, 0, 7);
 
     return util::make_string(
